+++ Plugins/Object/Object.hpp
@@ -62,6 +62,8 @@ private:
     ArgumentStack DoSpellLevelAbsorption    (ArgumentStack&& args);
     ArgumentStack SetHasInventory           (ArgumentStack&& args);
     ArgumentStack GetCurrentAnimation       (ArgumentStack&& args);
+    ArgumentStack GetAILevel                (ArgumentStack&& args);
+    ArgumentStack SetAILevel                (ArgumentStack&& args);
 
     CNWSObject *object(ArgumentStack& args);
 };